+++ web/src/beta/features/Editor/Widgets/WASToolsPanel/index.tsx
@@ -6,7 +6,7 @@ import { Panel } from "@reearth/beta/ui/layout";
 import { useT } from "@reearth/services/i18n";
 import { styled } from "@reearth/services/theme";
 
-import { useWidgetsPage } from "../widgetsPageContext";
+import { useWidgetsPage } from "../context";
 
 import Devices from "./Devices";
