+++ web/src/beta/features/Editor/hooks/useEditorVisualizer.ts
@@ -16,6 +16,7 @@ export default () => {
 
   // Visualizer Size
   const [visualizerSize, setVisualizerSize] = useState({ width: 0, height: 0, left: 0, top: 0 });
+  const isVisualizerResizing = useRef(false);
 
   const handleVisuzlierResize = useCallback((size: AreaSize) => {
     setVisualizerSize({
@@ -23,6 +24,10 @@ export default () => {
       top: size.top + 1,
       width: size.width,
       height: size.height,
+    });
+    isVisualizerResizing.current = true;
+    requestAnimationFrame(() => {
+      isVisualizerResizing.current = false;
     });
   }, []);
 
@@ -48,6 +53,7 @@ export default () => {
     handleIsVisualizerUpdate,
     visualizerSize,
     handleVisuzlierResize,
+    isVisualizerResizing,
     currentCamera,
     handleCameraUpdate,
     handleFlyTo,
